+++ lib/commands/bidi/models.ts
@@ -4,7 +4,7 @@ import type {
   BiDiLogLevel,
   NetworkMonitorBiDiEvent,
 } from './types';
-import {NATIVE_WIN} from '../../utils';
+import {NATIVE_WIN} from '../constants';
 import {CONTEXT_UPDATED_EVENT, LOG_ENTRY_ADDED_EVENT, NETWORK_MONITOR_EVENT} from './constants';
 import type {LogEntry} from '../types';
 import _ from 'lodash';
