+++ lib/commands/certificate.ts
@@ -8,7 +8,8 @@ import http from 'node:http';
 import {exec} from 'teen_process';
 import {findAPortNotInUse, checkPortStatus} from 'portscanner';
 import {CertificateClient} from '../device/certificate-client';
-import {isIos18OrNewer, requireRealDevice} from '../utils';
+import {isIos18OrNewer} from '../utils';
+import {requireRealDevice} from './guards';
 import type {Simulator} from 'appium-ios-simulator';
 import type {XCUITestDriver} from '../driver';
 import type {CertificateList} from './types';
