+++ lib/commands/condition.ts
@@ -1,7 +1,7 @@
 import type {XCUITestDriver} from '../driver';
 import {createConditionInducer} from '../device/condition-inducer-client';
 import type {Condition} from '../types';
-import {requireRealDevice} from '../utils';
+import {requireRealDevice} from './guards';
 
 /**
  * Get all available ConditionInducer configuration information, which can be used with
