+++ lib/commands/constants.ts
@@ -1,0 +1,9 @@
+/** Capability value that selects an available simulator automatically. */
+export const UDID_AUTO = 'auto';
+
+/** Native (non-webview) context identifier. */
+export const NATIVE_WIN = 'NATIVE_APP';
+
+export const APP_EXT = '.app';
+export const IPA_EXT = '.ipa';
+export const SUPPORTED_EXTENSIONS = [IPA_EXT, APP_EXT];
