+++ lib/commands/gesture.ts
@@ -1,7 +1,7 @@
 import {errors} from 'appium/driver';
 import {util} from 'appium/support';
 import _ from 'lodash';
-import {requireSimulator} from '../utils';
+import {requireSimulator} from './guards';
 import type {XCUITestDriver} from '../driver';
 import type {ActionSequence, Element} from '@appium/types';
 import type {Direction} from './types';
