+++ lib/commands/localization.ts
@@ -1,4 +1,4 @@
-import {requireSimulator} from '../utils';
+import {requireSimulator} from './guards';
 import type {XCUITestDriver} from '../driver';
 import type {KeyboardOptions, LanguageOptions, LocaleOptions} from './types';
