+++ lib/commands/log.ts
@@ -7,7 +7,8 @@ import WebSocket from 'ws';
 import {SafariConsoleLog} from '../device/log/safari-console-log';
 import {SafariNetworkLog} from '../device/log/safari-network-log';
 import {toLogEntry} from '../device/log/helpers';
-import {NATIVE_WIN, isIos18OrNewer} from '../utils';
+import {isIos18OrNewer} from '../utils';
+import {NATIVE_WIN} from './constants';
 import {BIDI_EVENT_NAME} from './bidi/constants';
 import {makeLogEntryAddedEvent} from './bidi/models';
 import type {XCUITestDriver} from '../driver';
