+++ lib/commands/network-monitor.ts
@@ -1,6 +1,7 @@
 import {errors} from 'appium/driver';
 import {NetworkMonitorSession} from '../device/network-monitor-session';
-import {isIos18OrNewer, requireRealDevice} from '../utils';
+import {isIos18OrNewer} from '../utils';
+import {requireRealDevice} from './guards';
 import type {XCUITestDriver} from '../driver';
 
 /**
