+++ lib/commands/pasteboard.ts
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import {requireSimulator} from '../utils';
+import {requireSimulator} from './guards';
 import type {XCUITestDriver} from '../driver';
 
 /**
