+++ lib/commands/web.ts
@@ -3,7 +3,8 @@ import {timing, util} from 'appium/support';
 import {retryInterval} from 'asyncbox';
 import _ from 'lodash';
 import {setTimeout as delay} from 'node:timers/promises';
-import {withTimeout, TimeoutError, requireSimulator} from '../utils';
+import {withTimeout, TimeoutError} from '../utils';
+import {requireSimulator} from './guards';
 import type {XCUITestDriver} from '../driver';
 import type {Element, Cookie, Size, Position, Rect} from '@appium/types';
 import type {AtomsElement} from './types';
