+++ lib/utils/index.ts
@@ -1,7 +1,7 @@
-export * from './app';
+export * from './platform';
 export * from './async';
-export * from './driver';
 export * from './network';
 export * from './runtime';
 export * from './validation';
 export * from './xcode';
+export * from './app';
